requirements.txt
fastapi
uvicorn
SQLAlchemy
python-docx-template
lxml
beautifulsoup4
python-multipart
openai
python-dotenv